Move gcs & wasb task handlers to their respective provider packages (#9714)
