+++ airflow/providers/microsoft/azure/log/__init__.py
@@ -1,4 +1,3 @@
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -15,15 +14,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""
-This module is deprecated. Please use `airflow.providers.microsoft.azure.log.wasb_task_handler`.
-"""
-import warnings
-
-# pylint: disable=unused-import
-from airflow.providers.microsoft.azure.log.wasb_task_handler import WasbTaskHandler  # noqa
-
-warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.microsoft.azure.log.wasb_task_handler`.",
-    DeprecationWarning, stacklevel=2
-)
