+++ airflow/utils/log/gcs_task_handler.py
@@ -15,168 +15,15 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-import os
-from urllib.parse import urlparse
+"""
+This module is deprecated. Please use `airflow.providers.google.cloud.log.gcs_task_handler`.
+"""
+import warnings
 
-from cached_property import cached_property
+# pylint: disable=unused-import
+from airflow.providers.google.cloud.log.gcs_task_handler import GCSTaskHandler  # noqa
 
-from airflow.configuration import conf
-from airflow.exceptions import AirflowException
-from airflow.utils.log.file_task_handler import FileTaskHandler
-from airflow.utils.log.logging_mixin import LoggingMixin
-
-
-class GCSTaskHandler(FileTaskHandler, LoggingMixin):
-    """
-    GCSTaskHandler is a python log handler that handles and reads
-    task instance logs. It extends airflow FileTaskHandler and
-    uploads to and reads from GCS remote storage. Upon log reading
-    failure, it reads from host machine's local disk.
-    """
-    def __init__(self, base_log_folder, gcs_log_folder, filename_template):
-        super().__init__(base_log_folder, filename_template)
-        self.remote_base = gcs_log_folder
-        self.log_relative_path = ''
-        self._hook = None
-        self.closed = False
-        self.upload_on_close = True
-
-    @cached_property
-    def hook(self):
-        """
-        Returns GCS hook.
-        """
-        remote_conn_id = conf.get('logging', 'REMOTE_LOG_CONN_ID')
-        try:
-            from airflow.providers.google.cloud.hooks.gcs import GCSHook
-            return GCSHook(
-                google_cloud_storage_conn_id=remote_conn_id
-            )
-        except Exception as e:  # pylint: disable=broad-except
-            self.log.error(
-                'Could not create a GoogleCloudStorageHook with connection id '
-                '"%s". %s\n\nPlease make sure that airflow[gcp] is installed '
-                'and the GCS connection exists.', remote_conn_id, str(e)
-            )
-
-    def set_context(self, ti):
-        super().set_context(ti)
-        # Log relative path is used to construct local and remote
-        # log path to upload log files into GCS and read from the
-        # remote location.
-        self.log_relative_path = self._render_filename(ti, ti.try_number)
-        self.upload_on_close = not ti.raw
-
-    def close(self):
-        """
-        Close and upload local log file to remote storage GCS.
-        """
-        # When application exit, system shuts down all handlers by
-        # calling close method. Here we check if logger is already
-        # closed to prevent uploading the log to remote storage multiple
-        # times when `logging.shutdown` is called.
-        if self.closed:
-            return
-
-        super().close()
-
-        if not self.upload_on_close:
-            return
-
-        local_loc = os.path.join(self.local_base, self.log_relative_path)
-        remote_loc = os.path.join(self.remote_base, self.log_relative_path)
-        if os.path.exists(local_loc):
-            # read log and remove old logs to get just the latest additions
-            with open(local_loc, 'r') as logfile:
-                log = logfile.read()
-            self.gcs_write(log, remote_loc)
-
-        # Mark closed so we don't double write if close is called twice
-        self.closed = True
-
-    def _read(self, ti, try_number, metadata=None):
-        """
-        Read logs of given task instance and try_number from GCS.
-        If failed, read the log from task instance host machine.
-
-        :param ti: task instance object
-        :param try_number: task instance try_number to read logs from
-        :param metadata: log metadata,
-                         can be used for steaming log reading and auto-tailing.
-        """
-        # Explicitly getting log relative path is necessary as the given
-        # task instance might be different than task instance passed in
-        # in set_context method.
-        log_relative_path = self._render_filename(ti, try_number)
-        remote_loc = os.path.join(self.remote_base, log_relative_path)
-
-        try:
-            remote_log = self.gcs_read(remote_loc)
-            log = '*** Reading remote log from {}.\n{}\n'.format(
-                remote_loc, remote_log)
-            return log, {'end_of_log': True}
-        except Exception as e:  # pylint: disable=broad-except
-            log = '*** Unable to read remote log from {}\n*** {}\n\n'.format(
-                remote_loc, str(e))
-            self.log.error(log)
-            local_log, metadata = super()._read(ti, try_number)
-            log += local_log
-            return log, metadata
-
-    def gcs_read(self, remote_log_location):
-        """
-        Returns the log found at the remote_log_location.
-
-        :param remote_log_location: the log's location in remote storage
-        :type remote_log_location: str (path)
-        """
-        bkt, blob = self.parse_gcs_url(remote_log_location)
-        return self.hook.download(bkt, blob).decode('utf-8')
-
-    def gcs_write(self, log, remote_log_location, append=True):
-        """
-        Writes the log to the remote_log_location. Fails silently if no hook
-        was created.
-
-        :param log: the log to write to the remote_log_location
-        :type log: str
-        :param remote_log_location: the log's location in remote storage
-        :type remote_log_location: str (path)
-        :param append: if False, any existing log file is overwritten. If True,
-            the new log is appended to any existing logs.
-        :type append: bool
-        """
-        if append:
-            try:
-                old_log = self.gcs_read(remote_log_location)
-                log = '\n'.join([old_log, log]) if old_log else log
-            except Exception as e:  # pylint: disable=broad-except
-                if not hasattr(e, 'resp') or e.resp.get('status') != '404':  # pylint: disable=no-member
-                    log = '*** Previous log discarded: {}\n\n'.format(str(e)) + log
-
-        try:
-            bkt, blob = self.parse_gcs_url(remote_log_location)
-            from tempfile import NamedTemporaryFile
-            with NamedTemporaryFile(mode='w+') as tmpfile:
-                tmpfile.write(log)
-                # Force the file to be flushed, since we're doing the
-                # upload from within the file context (it hasn't been
-                # closed).
-                tmpfile.flush()
-                self.hook.upload(bkt, blob, tmpfile.name)
-        except Exception as e:  # pylint: disable=broad-except
-            self.log.error('Could not write logs to %s: %s', remote_log_location, e)
-
-    @staticmethod
-    def parse_gcs_url(gsurl):
-        """
-        Given a Google Cloud Storage URL (gs://<bucket>/<blob>), returns a
-        tuple containing the corresponding bucket and blob.
-        """
-        parsed_url = urlparse(gsurl)
-        if not parsed_url.netloc:
-            raise AirflowException('Please provide a bucket name')
-        else:
-            bucket = parsed_url.netloc
-            blob = parsed_url.path.strip('/')
-            return bucket, blob
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.google.cloud.log.gcs_task_handler`.",
+    DeprecationWarning, stacklevel=2
+)
